Push proof dependencies from VerifyPendingRegistrationsTask to proof package
